[service-utils] Add kafka producer to emit raw messages
