+++ .changeset/thirty-monkeys-promise.md
@@ -1,0 +1,5 @@
+---
+"@thirdweb-dev/service-utils": patch
+---
+
+Add raw kafka producer
